pyproject.toml
[project]
name = "autotestgen-tests"
version = "0.1.0"
description = ""
authors = [
    {name = "Armen Asatryan",email = "armen.asatryan@inconceptlabs.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client (>=5.2.0,<6.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
